pyproject.toml
[project]
name = "credit-assessment"
version = "0.1.0"
license = "LicenseRef-Proprietary"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "fastapi>=0.110,<1",
    "uvicorn>=0.25,<1",
    "slowapi>=0.1,<1",
    "structlog>=24.0,<25",
    "sqlalchemy[asyncio]>=2.0,<3",
    "aiosqlite>=0.20,<1",
    "alembic>=1.13,<2",
    "PyJWT>=2.8,<3",
    "cryptography>=46.0.5,<48",
    "bcrypt>=4.0,<5",
    "passlib[bcrypt]>=1.7,<2",
    "prometheus-fastapi-instrumentator>=7.0,<8",
    "sentry-sdk[fastapi]>=2.0,<3",
    "email-validator>=2.0,<3",
    "stripe>=8.0,<10",
    "httpx>=0.27,<1",

    "gunicorn>=22.0,<24",
    "redis>=5.0,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10",
    "pytest-asyncio>=0.23,<2",
    "pytest-cov>=5.0,<8",
    "httpx>=0.27,<1",
    "ruff>=0.4,<1",
    "coverage>=7.0,<8",
    "locust>=2.0,<3",
    "flask>=3.1.3,<4",
    "werkzeug>=3.1.6,<4",
    "pyyaml>=6.0,<7",
    "pre-commit>=4.0,<5",
]

[tool.pytest.ini_options]
testpaths = ["src"]
pythonpath = ["src"]
filterwarnings = [
    "ignore:Exception in thread.*_connection_worker_thread:pytest.PytestUnhandledThreadExceptionWarning",
]

[tool.coverage.run]
source = ["src/modules/credit"]
omit = ["*/tests/*"]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
fail_under = 100
show_missing = true
